Check django-admin executable instead of findfile().
Potentially fixes #56. We were doing a findfile for manage.py to try and identify if we're testing a django app. Instead, this commit changes the test for django to see if the django-admin executable exists in the virtualenv. We still do findfile() later on to identify the path to the manage.py file, but hopefully this will reduce the number of false positives.